Force contents of IFRAMEs to be ignored since they can't be written anyway
Following the logic set for noscript tags, iframe tag contents should
be ignored. Per documentation, they cannot be written to the page
anyway. This causes a problem for DART ads which include a script and
noscript tag inside an iframe, and caused the contained script tag to
be buffered, calling the ad twice.